[lsra]: Fix nested live ranges due to `finishLoop`
In certain edge cases, `finishLoop` would not insert `LsraEnd`s for
active variables resulting in nested live ranges which manifested as an
infinite loop in the register allocator (see discussion here [1]).

This commit fixes this bug by keeping track of whether a variable will
be made active in the future or not (i.e. if it will have a live range).
If a variable will have a live range in the future, then `finishLoop`
will make sure to insert `LsraEnd`s for them to avoid the above bug.

[1]: https://github.com/titzer/virgil/pull/308